Use levenshtein score to approximate moves.
One of the most common sources of diffs we are seeing comparing the filesystems of APKs is when there is an ABI change, but not an actual capability change.  This shows up as a potentially noisy Deleted / Added with the full capability listing.

This change adds a post-pass to the reporting computation that walks the Added/Deleted file lists, and if two paths are a >90% match, then we treat them as a "move" and combine their reports in a similar fashion to modifications.

This also incorporates the notion of moves into the report and adjusts some of the rendering to surface these as moves with the similarity score (so we can tune this).

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>