Only calculate moves for shared objects
Turns out this is really slow and for large numbers of files this ends
up taking hours. Since this is just a heuristic anyway and the only
given examples where this is useful were shared objects, add a dumb
little check for ".so." in the filename to avoid global warming.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>