mal: don't exit on extraction errors by default
malcontent defaults to exiting on extraction errors by default, which
ends up being something a lot of people trip over when trying to figure
out what's being detected in an artifact. Fix that by defaulting to not
exiting on extraction errors (with the prior change, malcontent will
still report them).

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>