Persist compiled rules locally rather than in-memory (#1122)
We currently persist compiled rules in-memory which only works for the duration of a single mal invocation. In cases where we want to run successive mal scans (usually when looping over specific files or directories to produce per-scan result files), we run into rule compilation overhead which takes at least several seconds per run which can be extremely slow when done dozens of times.

This PR instead stores rules locally in the user's cache directory using the compiled rule hash. This file is read from each time mal is run and will only be recreated if it does not exist.

I also added tests and benchmarks so we can validate this works and is faster (which it is by a factor of 10-12x)

Signed-off-by: egibs <20933572+egibs@users.noreply.github.com>